map svg colour improvements.
